Update to 1.6
